pyproject.toml
[project]
name = "unifiedcli"
version = "0.1.0"
description = "Multi-agent AI execution system that converts natural language goals into deployed web demos"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.0",
    "click>=8.0",
    "pyyaml>=6.0",
    "anthropic>=0.40",
    "openai>=1.0",
    "google-genai>=1.0",
]

[project.scripts]
unifiedcli = "unifiedcli.cli:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/unifiedcli"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
]
